inline func inside scope.c to scope.h

+++ Inc/ui/gui.h
@@ -85,7 +85,7 @@ void updateMeasData(void);
 
 
 extern uint16_t ADC_buffer[][320];
-extern float ADC_Vmax;
+extern int ADC_Vmax;
 static bool UI_data_ready;
 
 #endif /* _GUI_H_ */
